windows: Avoid heap mismatch when linked with multiple msvcr<X>.dll
libusb-1.0.dll is compiled with WDK 7 that links it with the original
msvcrt.dll. If brickd.exe is also linked with msvcrt.dll then everything
is okay, there is only one C runtime loaded and one heap in use.

But if brickd.exe is compiled using MSVC then it is linked with the
msvcr<X>.dll version specific to that MSVC compiler. Now there are two
different C runtimes loaded and two heaps in use. If libusb allocates
memory it is allocated on the msvcrt.dll heap, but when brickd.exe
tries to free such memory it will try to do that on the msvcr<X>.dll
heap. This doesn't work and results in a crash.

The problematic function that triggers this is libusb_get_pollfds. It
allocates memory for the list of pollfds and the caller (brickd.exe)
is supposed to free it afterwards using the normal free function. This
results in the mentioned heap mismatch.

Add a libusb_free function to libusb that allows brickd.exe to free
the memory allocated by libusb_get_pollfds on the correct heap.