Add option to use libudev via dlopen
This is necessary to be able to have one brickd binary that can handle
both SONAMEs of libudev: libudev.so.0 and libudev.so.1.

Debian currently ships libudev.so.0, but Ubuntu ships libudev.so.1 since
release 13.04.

By default brickd is linked normaly to libudev. To enable dlopen handling
WITH_LIBUDEV_DLOPEN=yes has to be spedified for make.