Use libusb hotplug support if available on Linux and Mac OSX
Since 1.0.16 libusb supports hotplug on Linux and Mac OSX. This can
result in a race between libusb and brickd handling hotplug events.
libusb tries to avoid this, but it can still happen. See usb_posix.c
for a more detailed explanation.

To avoid this race detect at runtime if libusb supports hotplug. If
it does let it handle hotplug and do no hotplug detection in brickd.
If it doesn't do hotplug detection in brickd as before.

If libusb 1.0.16 is detected at compile time then don't include libudev
and IOKit handling at all and rely on libusb. The Debian and Mac OSX
packages will always be build with libudev and IOKit. In combination
with the runtime detection of libusb hotplug this results in a brickd
binary that can work with any libusb version since 1.0.8.