Improve handling of request packets with invalid length in the header
Before, packets with a length < 8 in the header were detected, the
length value was ignored and the entire header was skipped, but the
client was not disconnected. Now such packets result in diconnecting
the client.

Before, packets with a length > 80 in the header were not detected.
In the end the client_handle_receive function tried to receive the
specified amount of data but failed in doing so, because the buffer
to store the data is only 80 bytes in size and socket_receive was
called with length 0 making it return 0. This was then erroneously
interpreted as a disconnect by peer in client_handle_receive. Now the
length in the packet header is checked before trying to receive the
payload and the client is disconnected if an invalid header is received.