sha1: Fix warning about possible loss of data
This warning is due to an assignment from size_t to unsigned char.

Change the type of the count variable from unsigned char to site_t to
avoid this warning.

This change is safe because the code doesn't depend on the size of
the sha1_ctxt struct and because the count variable will never store
a value greater than 127.