websocket: Fix parsing of multiple frames from a single buffer
If multiple frames arrived in a single buffer then the recursive
parsing logic accumulated the payload length incorrectly. Instead
of accumulating the payload length from each parsed frame the
remaining buffer size was accumulated. This resulted in returning
a large payload length that violated the contract of the socket
receive function as the return value could be larger than the
provided buffer length. In the end this triggered a buffer overflow
in the code calling the socket receive function.