websocket: Fix buffer handling in case frame arrives fragmented
If the websocket header is fragmented then upon receiving the
remaining part of the header the code would incorrectly skip a
full websocket header instead of just the remaining part. This
can result in a buffer overflow and misinterpretation of the
data behind the header.

Fix this by skipping only the remaining header part from the data.