Replace SHA1 implementation that has more robust endianess handling
The old one depended on the BYTE_ORDER macro provided by the compiler.
Those macros are not provided by GCC with C99 mode enabled. In this
case the old code broke because it didn't properly check for the
existence of those macros.

The new SHA1 implementation handles all its endianess needs by htonl().