Don't give up if claiming the USB interface of a Brick fails at first try
There seems to be a timing issue that makes the uevent for a USB device
arrival arrive at brickd before the actual USBFS entry for this device is
working. This makes the claiming of the interface fail if it is done too
fast after the uvent arrival.

Add a sleep/retry loop with 10 tries to handle this situation better.

This problem was found on the RED Brick.