Fix event source removal for events set to -1
If events is set to -1 the intended semantic is to remove all event
sources with a matching handle and type. But the previous logic was
that there could only be one event source with a given handle and type
combination. Therefore, the remove function included a shortcut to
exit as soon as the first matching event source was found and removed.

Remove this shortcut in case events is set to -1 to find and remove
all matching event sources.