Revert "Add option to not disconnect client on error"
This reverts commit f6b55ba58ac224d0d420ebee3471a3f1af2ee524.

This was meant for the RED Brick USB gadget interface. Brick Daemon should
not disconnect on it in case of an error, but try to keep the connection
alive by ignoring errors and trying to resync on invalid packets.

A better approach to this problem would be to disconnect/reconnect and to
make /dev/g_red_brick_data ensure that the data starts at a packet border
each time the device file is opened by brickd. g_red_brick can do this
because it knows the packet borders. This allows brickd to resync with the
USB packet stream just by reopening /dev/g_red_brick_data.