Fix corner cases in pending request handing
Don't add requests as pending if they are dropped, because they came
from an non-authenticated client.

Ensure that pending requests are removed from the list in all cases
if a matching response arrives. Before, pending requests were not
correctly removed from the list if they arrived after the client got
already disconnected or if the client was not authenticated.