Better avoid broadcasting unexpected responses
Before, the pending requests list per client was discarded on disconnect.
This resulted in broadcasting the responses that arrived after a client
that was expecteding them got disconnected. Such responses might then be
misrouted to the wrong client.

Instead of directly discarding the pending request list on disconnect of
a client, create a zombie that inherits then pending request list and
wait for 1 sec for the expected responses. This avoids broadcasting such
reponses.