windows: Improve RED Brick unplug detection and error handling
If a RED Brick is unplugged from a Windows PC then the one USB read
transfer gets stalled but all the others return successfully, but with
garbage data. It's unclear where this data comes from, if the cause is
on the RED Brick or the Windows side. Also resubmitting the successful
read transfers fails with I/O errors. The problem with this is that it
can spill 20 or more error messages into the log output for every RED
Brick unplug action.

The device removal event cannot be used to fix this, as it arrives too
late. Also just lowering USB transfer submission errors and packet
errors due to garbage data from error to debug level (to hide them from
the default log output) is not good, as it would also hide real errors.

To fix this an USB read transfer stall for a RED Brick is interpreted
as the first indication of an upcoming device removal. As consequence all
USB transfer handling for the affected device is stopped: no incoming
data is handled anymore, no outgoing data is send anymore, no read
transfers are resubmitted. This allows to avoid all the unwanted error
messages.