Assume that the USB device gets disconnected if transfer is cancelled
Or if the transfer fails with a no-device error. This stops transfer
submission for the USB device. This help to deal with the UWP problem
that the 8th read transfer fails with a CRC error and gets resubmitted
while the device is actually already removed. Transfer 1 to 7 failed
with a no-device error before. The resubmission of the 8th transfer
then makes the UWP app crash in a way that even the Visual Studio
debugger doesn't catch. Stopping the resubmission of the 8th transfer
helps to work around this problem.