Reopen USB device to recover from a stalled transfer
Before, a stalled transfer would not be submitted again and after all
transfers got stalled no communication with the USB device can occur
anymore.

In most cases a transfer will stall as a result of unplugging the USB
device. Use a 1 second timer to delay the reopening to avoid trying to
reopen an already unplugged USB device.