android: Switch from Android USB API to Linux usbfs
The Android USB API is not well design and way to sparse. It is not easy
to reimplement the libusb API with it because of the different event
handling strategies.

Instead get the usbfs file descriptor from the Android USB API and then
use ioctl calls to reeimplement the libusb API. This works way better.