uwp: Add experimental HAT (Zero) Brick support for Windows 10 IoT Core
It's experimental because Windows 10 IoT Core doesn't provide HAT
detection for the Raspberry Pi and it also doesn't allow to access the
I2C interface used for HAT detection to solve this in brickd itself.

This means that brickd cannot detect if and which HAT is connected.
Thereforem HAT support cannot be enabled by default and the define
BRICKD_WITH_BRICKLET has to be defined manually to enable it.