Add bricklet_stack_bcm2835.
This bricklet_stack implementation can be used instead of
bricklet_stack_linux on a Raspberry Pi by compiling with
WITH_BCM2835=yes.

The bcm2835 library communicates directly with the BCM2835 chip
on the Raspberry Pi (via /dev/mem). This results in much better
performance. Also changes in Raspbian's device tree don't affect
this implementation, so it works on newer kernels even if the
HAT's EEPROM emulation does not work (f.e. if the HAT is not
flashed at all or in the bootloader mode)