linux: Use correct core_freq to derive SPI frequency in BCM2835 backend
Before, the BCM2835 backend assumed core_freq to be fixed at 250 MHz.
This is only true for Raspberry Pi 1 and 2. All other Raspberry Pis
have a core_freq of 400 Mhz or higher, but might scale down the core
clock to 250 MHz assuming the OS is configured that way. This results
in a SPI frequency that might be way too high for the Bricklets to
understand, depending on the OS config.

Now, use the vcgencmd interface to get the actual core_freq value to
derive the correct SPI frequency from.