usb: Abort delayed stall recovery if device was removed in the meantime
There might be situations where the OS (Windows in this case) might
take longer than one second between an USB device being removed and
reporting this event to brickd. Brickd first sees a read transfer
failing with a stall condition. This starts the one seconds recovery
timer that triggers before the device removal event arrives. Therefore,
brickd tries to recover from the stall. This fails, because the device
is actually already removed. As a last ditch effort brickd tries to
reopen the USB device, which also fails, resulting in not sending the
enumerate-disconnected callbacks for the now disconnected USB device.

Fix this by paying attention to the expecting_removal flag and the
LIBUSB_ERROR_NO_DEVICE result during stall recovery. Also send the
enumerate-disconnected callbacks if a USB device cannot be reopened.