Use one libusb context for everything
Historically one context for used for enumeration only and each USBStack
had its own context. This was done to limit the impact of libusb leaking
USB transfers. After increasing the libusb requirement to 1.0.20 this is
not necessary anymore.

This is done in preparation for using libusb 1.0.24 on Windows with its
reworked inner event handling. For libusb 1.0.24 brickd will stop
accessing libusb's inner event handling. With the lack of pollable FDs
on Windows a libusb event handling thread has to be used. This is way
simpler if there is only one libusb context.