Do not release a destroyed semaphore
The thread-running variable was used in the cleanup path to determine
if the thread ends due to an error. But the thread-running variable is
also false in case the thread is ending normally.

Use a dedicated variable to keep track if the handshake semaphore was
not released yet in the cleanup path.