Add missing error event registration for libusb FDs
Without this brickd runs in a tight loop from USB device unplug until
unplug notification. This happens because the FD for a disconnected
USB device is in error state, but the event loop will not call the
libusb event handler as libusb did not register for the error event,
but will handle it.