Delay initial USB device scan to avoid crash
Due to the way libusb handles events internally a call to
libusb_get_string_descriptor_ascii (as part of the USB stack
creation) drives the libusb event handler and can result in
receiving USB responses from Bricks while brickd is still
in initalization phase and not all subsystems are initialized
yet.

In this case a USB response is received this way and the USB
stack tries to dispatch it to the network before the network
subsystem is initialized, resulting in a crash.

Avoid this by starting the initial USB device scan after all
subsystems